cheapdeck/pyproject.toml
[tool.poetry]
name = "cheapdeck"
version = "0.1.0"
description = "Stream deck using a numpad to control OBS over websockets"
authors = ["Erick Navarro <erick@navarro.io>"]

[tool.poetry.dependencies]
python = "^3.10"
websockets = "^10.3"
pyusb = "^1.2.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
